Move AutoDrainingCounter to top-level into concurrent package
* Move the AutoDrainingCounter out of AutoDrainingCounterTest and
  to a new package, 'concurrent', to indicate it is related to and uses
  various concurrent data structures (AtomicInteger, AtomicBoolean) as
  well as Java concurrency primitives, specifically it uses a
  ScheduledExecutorService to perform the drain operation on a recurring
  basis.
